implements neo4j as a embedded database

+++ src/main/kotlin/org/teknux/webapp/controller/ClockActionController.kt
@@ -14,7 +14,7 @@ class ClockActionController {
     private lateinit var storeService: IStoreService
 
     @GetMapping
-    fun getAll(@RequestParam userId: Int?): List<ClockAction> {
+    fun getAll(@RequestParam userId: Long?): List<ClockAction> {
         return userId?.let {
             storeService.getActions(it)
         } ?: storeService.getActions()
